requirements.txt
flask>=2.0.0
flask-sqlalchemy>=3.0.0
flask-login>=0.5.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.0
gunicorn>=20.1.0  # For production deployment